Delete add.Scalar from schema
The comment on native_functions.yaml claims this is just to
solve a C++ API problem.  If this is true, we don't actually
need this in schema, and all I need to do is just manually
replicate the API in the C++ bindings.  Seeing if this indeed
works.

If this does work, it solves a performance problem on add.Scalar
where I introduced an extra redispatch.  If it doesn't work I'll
try something else.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: fb705e44add36d92168fd2eb688b578f54d47508
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49203